Add functional checker
This allows creating the checking function up front, amortising
the cost of parsing the subnets for each check.

My use case for use this library was as part of a webserver and I have
a list of subnets and I check every request. Performance is greatly
improved by caching the checking function up front.

This adds a negligible overhead to the existing one-shot process for
single subnets, a tiny increase in performance for the one-shot process
with multiple subnets and a large increase when repeatedly rechecking
the same subnet.

The performance tests have been updated as well to show the difference.